requirements-gpu.txt
backports.weakref==1.0rc1
bleach==1.5.0
cycler==0.10.0
graphviz==0.8.1
h5py==2.7.1
html5lib==0.9999999
Keras==2.0.8
Markdown==2.6.9
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
protobuf==3.4.0
pydot==1.2.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
scipy==1.0.0
six==1.11.0
tensorflow-gpu==1.2.1
Werkzeug==0.12.2
